+++ rocket.c
@@ -99,7 +99,7 @@ static void DoMatch(size_t distance, size_t length, size_t offset, void *user)
 	PutMatchByte(instance, offset_adjusted & 0xFF);
 }
 
-static unsigned int GetMatchCost(size_t distance, size_t length, void *user)
+static size_t GetMatchCost(size_t distance, size_t length, void *user)
 {
 	(void)distance;
 	(void)length;
