+++ saxman.c
@@ -97,7 +97,7 @@ static void DoMatch(size_t distance, size_t length, size_t offset, void *user)
 	PutMatchByte(instance, (((offset - 0x12) & 0xF00) >> 4) | (length - 3));
 }
 
-static unsigned int GetMatchCost(size_t distance, size_t length, void *user)
+static size_t GetMatchCost(size_t distance, size_t length, void *user)
 {
 	(void)distance;
 	(void)length;
